[COR] define function in global scope, and adjust arguments.
There are specific cases where defining functions within functions is a
good idea, but those are pretty rare exceptions, more reading, if
interested, here:
https://www.codespeedy.com/define-a-function-inside-a-function-in-python/

A number of the arguments also needed to be adjusted, because if defined
locally the function has automatic access to the last definition of
variable names.
This should be avoided (and is another reason to generally not
define functions within functions), as it makes the behaviour
(i.e. the actual values of the variables) implicit and thus less predictable.